# go/api_default.go
/*
 * Directorii Webhook API
 *
 * API for receiving webhooks related to a contractor.  To effectively use the API for receiving webhooks, follow these steps:  1. **Generate a Program Template**:    - Visit the Swagger Editor at [Swagger Editor](https://editor.swagger.io/).    - Use the \"Generate Server\" button to create a template that suits your needs.  2. **Customize Your Webhook Path** (optional):    - In the generated template, locate the route designated as `/lead-created-or-updated`.    - Replace this placeholder path with your specific webhook URL that will handle incoming requests.  3. **Set up on Directorii Webhook Page**:    - Once you have your customized webhook path, navigate to the Directorii Webhook page.    - Configure your webhook path according to the protocol, domain name and path of your application.  By following these steps, you can ensure that your API is properly set up to receive notifications.
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

import (
	"encoding/json"
	"fmt"
	"io"
	"log"
	"net/http"
)

func LeadCreatedOrUpdated(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")

	// Read the request body
	body, err := io.ReadAll(r.Body)
	if err != nil {
		log.Printf("Error reading body: %v", err)
		http.Error(w, "Unable to read request body", http.StatusBadRequest)
		return
	}
	defer r.Body.Close()

	// Deserialize the JSON into ContractorWebhookPayload
	var payload ContractorWebhookPayload
	if err := json.Unmarshal(body, &payload); err != nil {
		log.Printf("Error unmarshalling JSON: %v", err)
		http.Error(w, "Invalid JSON format", http.StatusBadRequest)
		return
	}

	// Serialize the ContractorWebhookPayload into JSON
	jsonData, err := json.Marshal(payload)
	if err != nil {
		fmt.Println("Error:", err)
		return
	}

	// Log the payload to console
	log.Printf("Received webhook payload: %+v", string(jsonData))

	w.WriteHeader(http.StatusOK)
}