# go/model_contractor_webhook_lead_info.go
/*
 * Directorii Webhook API
 *
 * API for receiving webhooks related to a contractor.  To effectively use the API for receiving webhooks, follow these steps:  1. **Generate a Program Template**:    - Visit the Swagger Editor at [Swagger Editor](https://editor.swagger.io/).    - Use the \"Generate Server\" button to create a template that suits your needs.  2. **Customize Your Webhook Path** (optional):    - In the generated template, locate the route designated as `/lead-created-or-updated`.    - Replace this placeholder path with your specific webhook URL that will handle incoming requests.  3. **Set up on Directorii Webhook Page**:    - Once you have your customized webhook path, navigate to the Directorii Webhook page.    - Configure your webhook path according to the protocol, domain name and path of your application.  By following these steps, you can ensure that your API is properly set up to receive notifications.
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger
import (
	"time"
)

type ContractorWebhookLeadInfo struct {
	// Unique identifier for the lead.
	Id int32 `json:"Id,omitempty"`
	// Status code of the lead.
	Status int32 `json:"Status,omitempty"`
	// Human-readable status name of the lead.
	StatusName string `json:"StatusName,omitempty"`
	// Creation timestamp of the lead.
	Created time.Time `json:"Created,omitempty"`
	// This endpoint allows users to submit leads through various sources, including website submissions, manual entries by contractors, and voucher-based leads. Each lead source is categorized for better tracking and analysis. 
	Source string `json:"Source,omitempty"`
	// First name of the lead.
	FirstName string `json:"FirstName,omitempty"`
	// Last name of the lead.
	LastName string `json:"LastName,omitempty"`
	// Service requested by the lead.
	Service string `json:"Service,omitempty"`
	// Address of the lead.
	Address string `json:"Address,omitempty"`
	// City of the lead.
	City string `json:"City,omitempty"`
	// Zip code of the lead's address.
	ZipCode string `json:"ZipCode,omitempty"`
	// Phone number of the lead.
	Phone string `json:"Phone,omitempty"`
	// Email address of the lead.
	Email string `json:"Email,omitempty"`
	// Additional information about the lead.
	Description string `json:"Description,omitempty"`
}